ai-soiree-python/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
httpx>=0.24.0
transformers
torch
fastapi
uvicorn[standard]